Split budget table section rows
