+++ apps/web/src/ui/tables/budget/table/sections/BudgetDerivedSection.tsx
@@ -18,11 +18,10 @@ import { formatFxAmount } from "@/ui/tables/fx/format";
 import {
   buildYearTotalStateClass,
   isNegativeValueOver,
-  renderColumnCells,
-  renderDerivedYearLoadingCells,
-  renderSubtotalYearLoadingCells,
   renderValueCells,
 } from "./shared";
+import { LiquidityRow } from "./derived/LiquidityRow";
+import { MetricRow } from "./derived/MetricRow";
 
 export type BudgetDerivedSectionProps = Readonly<{
   effectiveAllowlist: ReadonlySet<string> | null;
@@ -43,214 +42,6 @@ export type BudgetDerivedSectionProps = Readonly<{
   numberFormat: NumberFormat;
   openFxBreakdown: (month: string) => void;
 }>;
-
-type MetricRowProps = Readonly<{
-  label: string;
-  columnSequence: ReadonlyArray<ColumnEntry>;
-  currentMonth: string;
-  currentYear: string;
-  yearComputed: ReadonlyMap<string, YearTotalComputed>;
-  numberFormat: NumberFormat;
-  derivedMaskClass: string;
-  renderPastYear: (year: string, yearData: YearTotalComputed) => ReactElement;
-  renderFutureYear: (year: string, yearData: YearTotalComputed) => ReactElement;
-  renderCurrentYear: (year: string, yearData: YearTotalComputed) => ReactElement;
-  renderPastMonth: (month: string) => ReactElement;
-  renderFutureMonth: (month: string) => ReactElement;
-  renderCurrentMonth: (month: string) => ReactElement;
-  loadingKind: "subtotal" | "derived";
-  rowClassName: string;
-}>;
-
-type LiquidityRowProps = Readonly<{
-  liquidity: string;
-  columnSequence: ReadonlyArray<ColumnEntry>;
-  currentMonth: string;
-  currentYear: string;
-  yearComputed: ReadonlyMap<string, YearTotalComputed>;
-  numberFormat: NumberFormat;
-  derivedMaskClass: string;
-  mebByLiq: Readonly<Record<string, Readonly<Record<string, number>>>>;
-  projectedLiqBalances: ReadonlyMap<string, Readonly<Record<string, number>>>;
-}>;
-
-const MetricRow = (props: MetricRowProps): ReactElement => {
-  const {
-    label,
-    columnSequence,
-    currentMonth,
-    currentYear,
-    yearComputed,
-    rowClassName,
-    renderPastYear,
-    renderFutureYear,
-    renderCurrentYear,
-    renderPastMonth,
-    renderFutureMonth,
-    renderCurrentMonth,
-    loadingKind,
-  } = props;
-  const renderLoading = loadingKind === "subtotal" ? renderSubtotalYearLoadingCells : renderDerivedYearLoadingCells;
-
-  return (
-    <tr className={rowClassName}>
-      <td className={`${rowClassName === styles.directionRow ? styles.directionLabel : styles.categoryLabel} ${styles.stickyCol}`}>{label}</td>
-      <td className={styles.leftSpacer} />
-      {columnSequence.map((column) => {
-        const yearData = column.kind === "year-total" ? yearComputed.get(column.year) : undefined;
-        return renderColumnCells({
-          column,
-          currentMonth,
-          currentYear,
-          isYearLoading: column.kind === "year-total" && yearData === undefined,
-          renderYearLoading: (isCurrentYearValue) =>
-            renderLoading(column.kind === "year-total" ? column.year : "", isCurrentYearValue),
-          renderPastYear: () => {
-            if (column.kind !== "year-total" || yearData === undefined) {
-              return renderLoading(column.kind === "year-total" ? column.year : "", false);
-            }
-            return renderPastYear(column.year, yearData);
-          },
-          renderFutureYear: () => {
-            if (column.kind !== "year-total" || yearData === undefined) {
-              return renderLoading(column.kind === "year-total" ? column.year : "", false);
-            }
-            return renderFutureYear(column.year, yearData);
-          },
-          renderCurrentYear: () => {
-            if (column.kind !== "year-total" || yearData === undefined) {
-              return renderLoading(column.kind === "year-total" ? column.year : "", true);
-            }
-            return renderCurrentYear(column.year, yearData);
-          },
-          renderPastMonth: () => {
-            if (column.kind !== "month") {
-              return renderLoading("invalid", false);
-            }
-            return renderPastMonth(column.month);
-          },
-          renderFutureMonth: () => {
-            if (column.kind !== "month") {
-              return renderLoading("invalid", false);
-            }
-            return renderFutureMonth(column.month);
-          },
-          renderCurrentMonth: () => {
-            if (column.kind !== "month") {
-              return renderLoading("invalid", false);
-            }
-            return renderCurrentMonth(column.month);
-          },
-        });
-      })}
-    </tr>
-  );
-};
-
-const LiquidityRow = (props: LiquidityRowProps): ReactElement => {
-  const {
-    liquidity,
-    columnSequence,
-    currentMonth,
-    currentYear,
-    yearComputed,
-    numberFormat,
-    derivedMaskClass,
-    mebByLiq,
-    projectedLiqBalances,
-  } = props;
-  const { t } = useTranslation();
-
-  return (
-    <tr key={`bal-${liquidity}`} className={styles.categoryRow}>
-      <td className={`${styles.categoryLabel} ${styles.stickyCol}${derivedMaskClass}`}>
-        {t(`budget.liquidity${liquidity.charAt(0).toUpperCase()}${liquidity.slice(1)}`)}
-      </td>
-      <td className={styles.leftSpacer} />
-      {columnSequence.map((column) => {
-        const yearData = column.kind === "year-total" ? yearComputed.get(column.year) : undefined;
-        return renderColumnCells({
-          column,
-          currentMonth,
-          currentYear,
-          isYearLoading: column.kind === "year-total" && yearData === undefined,
-          renderYearLoading: (isCurrentYearValue) =>
-            renderDerivedYearLoadingCells(column.kind === "year-total" ? column.year : "", isCurrentYearValue),
-          renderPastYear: () => {
-            if (column.kind !== "year-total" || yearData === undefined) {
-              return renderDerivedYearLoadingCells(column.kind === "year-total" ? column.year : "", false);
-            }
-            return (
-              <td key={`total-${column.year}`} className={`${styles.cell} ${styles.yearTotal}${derivedMaskClass}`}>
-                {formatAmount(yearData.decemberBalancesByLiquidity[liquidity] ?? 0, numberFormat)}
-              </td>
-            );
-          },
-          renderFutureYear: () => {
-            if (column.kind !== "year-total" || yearData === undefined) {
-              return renderDerivedYearLoadingCells(column.kind === "year-total" ? column.year : "", false);
-            }
-            return (
-              <td key={`total-${column.year}`} className={`${styles.cell} ${styles.yearTotal}${derivedMaskClass}`}>
-                {formatAmount(yearData.decemberBalancesByLiquidityPlan[liquidity] ?? 0, numberFormat)}
-              </td>
-            );
-          },
-          renderCurrentYear: () => {
-            if (column.kind !== "year-total" || yearData === undefined) {
-              return renderDerivedYearLoadingCells(column.kind === "year-total" ? column.year : "", true);
-            }
-            return (
-              <Fragment key={`total-${column.year}`}>
-                <td className={`${styles.cell} ${styles.yearTotal}${derivedMaskClass}`}>
-                  {formatAmount(yearData.decemberBalancesByLiquidityPlan[liquidity] ?? 0, numberFormat)}
-                </td>
-                <td className={`${styles.cell} ${styles.yearTotal}${derivedMaskClass}`}>
-                  {formatAmount(yearData.decemberBalancesByLiquidity[liquidity] ?? 0, numberFormat)}
-                </td>
-              </Fragment>
-            );
-          },
-          renderPastMonth: () => {
-            if (column.kind !== "month") {
-              return renderDerivedYearLoadingCells("invalid", false);
-            }
-            return (
-              <td key={column.month} className={`${styles.cell}${derivedMaskClass}`}>
-                {formatAmount(mebByLiq[column.month]?.[liquidity] ?? 0, numberFormat)}
-              </td>
-            );
-          },
-          renderFutureMonth: () => {
-            if (column.kind !== "month") {
-              return renderDerivedYearLoadingCells("invalid", false);
-            }
-            return (
-              <td key={column.month} className={`${styles.cell}${derivedMaskClass}`}>
-                {formatAmount(projectedLiqBalances.get(column.month)?.[liquidity] ?? 0, numberFormat)}
-              </td>
-            );
-          },
-          renderCurrentMonth: () => {
-            if (column.kind !== "month") {
-              return renderDerivedYearLoadingCells("invalid", false);
-            }
-            return (
-              <Fragment key={column.month}>
-                <td className={`${styles.cell} ${styles.currentMonthPlan}${derivedMaskClass}`}>
-                  {formatAmount(projectedLiqBalances.get(column.month)?.[liquidity] ?? 0, numberFormat)}
-                </td>
-                <td className={`${styles.cell} ${styles.currentMonthActual}${derivedMaskClass}`}>
-                  {formatAmount(mebByLiq[column.month]?.[liquidity] ?? 0, numberFormat)}
-                </td>
-              </Fragment>
-            );
-          },
-        });
-      })}
-    </tr>
-  );
-};
 
 export const BudgetDerivedSection = (props: BudgetDerivedSectionProps): ReactElement => {
   const {
